+++ domain/domain-pos/src/main/java/domain/pos/menu/entity/v2/domain/Menu.java
@@ -1,4 +1,4 @@
-package domain.pos.menu.entity.v2;
+package domain.pos.menu.entity.v2.domain;
 
 import static java.util.Objects.*;
 
@@ -8,7 +8,7 @@ import com.exception.ErrorCode;
 import com.exception.ServiceException;
 import com.vo.AuditStamp;
 
-import domain.pos.menu.entity.v2.state.MenuInfoState;
+import domain.pos.menu.entity.v2.domain.state.MenuInfoState;
 import lombok.Builder;
 import lombok.Getter;
